Final/requirements.txt
PyGithub>=2.1.1
google-genai>=0.2.0
python-dotenv>=1.0.0
streamlit>=1.28.0

